requirements.txt
django-ninja==1.1.0
gunicorn
django-redis
redis
pika
